fixed the order of recipe files
