+++ .gitignore
@@ -94,3 +94,4 @@ Temporary Items
 /bin/Debug
 /Chapter12/chapter12_08_1/main.obj
 /Chapter12/chapter12_08_1/main.exe
+/out/build/x64-Debug
